+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ##### New features
 
+* License changed from MIT to [Apache-2.0](https://www.apache.org/licenses/LICENSE-2.0).
 * **Important!** The functionality of most of the methods included in TKSBrokerAPI v1.0 has been restored, except for opening a grid of orders and downloading historical data (to be added in future releases). Now all methods work with the new Open API: https://tinkoff.github.io/investAPI/swagger-ui/
 * **Important!** The TKSBrokerAPI module has been made as open-source project, further development continues there: https://github.com/Tim55667757/TKSBrokerAPI
 * **Important!** The Tinkoff Invest API version supported by the TKSBrokerAPI library is now v2 and not backwards compatible.
@@ -43,6 +44,7 @@
 * Moved from the `Overview()` method to separate methods `RequestPortfolio()`, `RequestPositions()`, `RequestPendingOrders` and `RequestStopOrders()` — operations on requesting a portfolio, open positions and user orders.
 * Using the `CloseOrders()` method, it is now possible to close both exchange pending and stop orders. It is enough to specify an ID or a list of IDs.
 * Added display of the difference in % between the previous and current closing price of the instrument in all tables where it is required.
+* [In progress] [#22](https://github.com/Tim55667757/TKSBrokerAPI/issues/22) `--open-trade` and `--open-order` keys were replaced with `--trade` and `--order` keys, because it's too long.
 
 ##### Bug fixes